+++ ecal/core/CMakeLists.txt
@@ -385,9 +385,9 @@ if(ECAL_CORE_SERVICE)
       src/service/ecal_service_client_v5.cpp
       src/service/ecal_service_client_v5_impl.cpp
       src/service/ecal_service_client_v5_impl.h
-      src/service/ecal_service_server.cpp
-      src/service/ecal_service_server_impl.cpp
-      src/service/ecal_service_server_impl.h
+      src/service/ecal_service_server_v5.cpp
+      src/service/ecal_service_server_v5_impl.cpp
+      src/service/ecal_service_server_v5_impl.h
       src/service/ecal_service_singleton_manager.cpp
       src/service/ecal_service_singleton_manager.h
   )
@@ -535,7 +535,7 @@ set(ecal_header_cmn
     include/ecal/ecal_process_severity.h
     include/ecal/ecal_registration.h
     include/ecal/ecal_publisher.h
-    include/ecal/ecal_server.h
+    include/ecal/ecal_server_v5.h
     include/ecal/ecal_service_info.h
     include/ecal/ecal_subscriber.h
     include/ecal/ecal_time.h
